+++ plone/app/vocabularies/__init__.py
@@ -1,13 +1,11 @@
 # -*- coding: utf-8 -*-
 from plone.app.vocabularies.interfaces import ISlicableVocabulary
 from plone.app.vocabularies.interfaces import IPermissiveVocabulary
+from six.moves import urllib
 from zope.interface import directlyProvides
 from zope.interface import implementer
 from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
-
-import urllib
-
 
 _token_parse_py3 = getattr(urllib, 'parse', None)
 _token_parse_py27 = lambda token: urllib.unquote_plus(token).decode('utf8')